+++ spec/serializing_spec.rb
@@ -6,18 +6,27 @@ RSpec.describe "serializing MemoWise" do # rubocop:disable RSpec/DescribeClass
       Class.new do
         prepend MemoWise
 
-        attr_reader :name, :name_upper_counter
+        attr_reader :name, :name_upper_counter, :hello_counter
 
         def initialize(name:)
           @name = name
           @name_upper_counter = 0
+          @hello_counter = 0
         end
 
+        # We test both 0- and 1-arity methods here because they use different
+        # storage schemes.
         def name_upper
           @name_upper_counter += 1
           name.upcase
         end
         memo_wise :name_upper
+
+        def hello(n) # rubocop:disable Naming/MethodParameterName
+          @hello_counter += 1
+          "Hello #{n} times, #{name}!"
+        end
+        memo_wise :hello
       end
     end
 
@@ -35,13 +44,18 @@ RSpec.describe "serializing MemoWise" do # rubocop:disable RSpec/DescribeClass
     it "dumps and loads memoized state" do
       obj1 = ClassForTest.new(name: "foo")
       obj1.name_upper
+      obj1.hello(3)
       obj2 = Marshal.load(Marshal.dump(obj1))
 
       expect { obj2.name_upper }.
         not_to change { obj2.name_upper_counter }.
         from(1)
+      expect { obj2.hello(3) }.
+        not_to change { obj2.hello_counter }.
+        from(1)
 
       expect(obj2.name_upper).to eq("FOO")
+      expect(obj2.hello(3)).to eq("Hello 3 times, foo!")
     end
   end
 end