+++ CHANGELOG.md
@@ -6,8 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
 ### Updated
-- (Nothing, yet)
+- Improve performance of single-argument methods by using an outer Array instead
+  of a Hash
 ### Fixed
 - (Nothing, yet)
 ### Breaking Changes